Updated VC++ 2005 and DevCpp projects, small cleanups and fixes
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@1031 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/common/FileDownload.cpp
@@ -264,7 +264,7 @@ byte CFileDownload::GetProgress()
 	Lock();
 	byte res = 0;
 	if (tHttp.GetDataLength() != 0)
-		res = MIN(100, tHttp.GetReceivedDataLen() * 100 / tHttp.GetDataLength());
+		res = (byte)MIN(100, tHttp.GetReceivedDataLen() * 100 / tHttp.GetDataLength());
 	Unlock();
 
 	return res;
